Rename terrainHandler.js to terrainRouter.js
